Move ConfigManager registration/dependency injection to RulesetContainer
