+++ osu.Game/OsuGame.cs
@@ -71,6 +71,7 @@ namespace osu.Game
         private OsuScreen screenStack;
 
         private VolumeControl volume;
+        private OnScreenDisplay onscreenDisplay;
 
         private Bindable<int> configRuleset;
         public Bindable<RulesetInfo> Ruleset = new Bindable<RulesetInfo>();
@@ -195,7 +196,7 @@ namespace osu.Game
             }, overlayContent.Add);
 
             loadComponentSingleFile(volume = new VolumeControl(), Add);
-            loadComponentSingleFile(new OnScreenDisplay(), Add);
+            loadComponentSingleFile(onscreenDisplay = new OnScreenDisplay(), Add);
 
             //overlay elements
             loadComponentSingleFile(direct = new DirectOverlay { Depth = -1 }, mainContent.Add);
@@ -232,6 +233,7 @@ namespace osu.Game
             forwardLoggedErrorsToNotifications();
 
             dependencies.Cache(settings);
+            dependencies.Cache(onscreenDisplay);
             dependencies.Cache(social);
             dependencies.Cache(direct);
             dependencies.Cache(chat);
